// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Page_User_: {
            /** Items */
            items: components["schemas"]["User"][];
            /** Limit */
            limit: number;
            /** Offset */
            offset: number;
            /** Total */
            total: number;
        };
// ...
        User: {
            /** Created At */
            created_at: string;
            /** Display Name */
            display_name: string;
            /** Email */
            email: string;
            /** Enabled */
            enabled: boolean;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Role */
            role: string;
            /** Updated At */
            updated_at: string;
            /** Version */
            version: number;
        };
        /** UserPatch */
        UserPatch: {
            /** Enabled */
            enabled?: boolean | null;
            /** Role */
            role?: ("member" | "admin") | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}